+++ sdk/storage/Azure.Storage.DataMovement.Blobs/src/BlobDestinationCheckpointDetails.cs
@@ -11,7 +11,7 @@ using Tags = System.Collections.Generic.IDictionary<string, string>;
 
 namespace Azure.Storage.DataMovement.Blobs
 {
-    internal class BlobDestinationCheckpointData : StorageResourceCheckpointData
+    internal class BlobDestinationCheckpointDetails : StorageResourceCheckpointDetails
     {
         public int Version;
 
@@ -66,7 +66,7 @@ namespace Azure.Storage.DataMovement.Blobs
 
         public override int Length => CalculateLength();
 
-        public BlobDestinationCheckpointData(
+        public BlobDestinationCheckpointDetails(
             DataTransferProperty<BlobType?> blobType,
             DataTransferProperty<string> contentType,
             DataTransferProperty<string> contentEncoding,
@@ -77,7 +77,7 @@ namespace Azure.Storage.DataMovement.Blobs
             DataTransferProperty<Metadata> metadata,
             DataTransferProperty<Tags> tags)
         {
-            Version = DataMovementBlobConstants.DestinationCheckpointData.SchemaVersion;
+            Version = DataMovementBlobConstants.DestinationCheckpointDetails.SchemaVersion;
             BlobType = blobType;
             PreserveBlobType = blobType?.Preserve ?? true;
             BlobTypeValue = blobType?.Value != default ? blobType.Value : default;
@@ -117,7 +117,7 @@ namespace Azure.Storage.DataMovement.Blobs
         {
             Argument.AssertNotNull(stream, nameof(stream));
 
-            int currentVariableLengthIndex = DataMovementBlobConstants.DestinationCheckpointData.VariableLengthStartIndex;
+            int currentVariableLengthIndex = DataMovementBlobConstants.DestinationCheckpointDetails.VariableLengthStartIndex;
             BinaryWriter writer = new BinaryWriter(stream);
 
             // Version
@@ -258,7 +258,7 @@ namespace Azure.Storage.DataMovement.Blobs
             }
         }
 
-        internal static BlobDestinationCheckpointData Deserialize(Stream stream)
+        internal static BlobDestinationCheckpointDetails Deserialize(Stream stream)
         {
             Argument.AssertNotNull(stream, nameof(stream));
 
@@ -266,7 +266,7 @@ namespace Azure.Storage.DataMovement.Blobs
 
             // Version
             int version = reader.ReadInt32();
-            if (version != DataMovementBlobConstants.DestinationCheckpointData.SchemaVersion)
+            if (version != DataMovementBlobConstants.DestinationCheckpointDetails.SchemaVersion)
             {
                 throw Errors.UnsupportedJobSchemaVersionHeader(version);
             }
@@ -376,7 +376,7 @@ namespace Azure.Storage.DataMovement.Blobs
                 tagsString = reader.ReadBytes(tagsLength).AsString();
             }
 
-            return new BlobDestinationCheckpointData(
+            return new BlobDestinationCheckpointDetails(
                 blobType: preserveBlobType ? new(preserveBlobType) : new(blobType),
                 contentType: preserveContentType ? new(preserveContentType) : new(contentType),
                 contentEncoding: preserveContentEncoding ? new(preserveContentEncoding): new(contentEncoding),
@@ -392,7 +392,7 @@ namespace Azure.Storage.DataMovement.Blobs
         {
             // Length is calculated based on whether the property is preserved.
             // If the property is preserved, the property's length is added to the total length.
-            int length = DataMovementBlobConstants.DestinationCheckpointData.VariableLengthStartIndex;
+            int length = DataMovementBlobConstants.DestinationCheckpointDetails.VariableLengthStartIndex;
             if (!PreserveContentType)
             {
                 length += ContentTypeBytes.Length;