+++ sdk/storage/Azure.Storage.DataMovement.Files.Shares/src/ShareFileSourceCheckpointDetails.cs
@@ -5,7 +5,7 @@ using System.IO;
 
 namespace Azure.Storage.DataMovement.Files.Shares
 {
-    internal class ShareFileSourceCheckpointData : StorageResourceCheckpointDataInternal
+    internal class ShareFileSourceCheckpointDetails : StorageResourceCheckpointDetailsInternal
     {
         public override int Length => 0;
 
@@ -13,6 +13,6 @@ namespace Azure.Storage.DataMovement.Files.Shares
         {
         }
 
-        internal static ShareFileSourceCheckpointData Deserialize(Stream stream) => new();
+        internal static ShareFileSourceCheckpointDetails Deserialize(Stream stream) => new();
     }
 }