+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/StorageResourceCheckpointDetailsInternal.cs
@@ -11,10 +11,10 @@ namespace Azure.Storage.DataMovement
     /// <summary>
     /// Middle class between the public type and implementation class.
     /// Gives internal hook methods to protected methods of
-    /// <see cref="StorageResourceCheckpointData"/>, allowing for internal
+    /// <see cref="StorageResourceCheckpointDetails"/>, allowing for internal
     /// package use as well as testing access.
     /// </summary>
-    internal abstract class StorageResourceCheckpointDataInternal : StorageResourceCheckpointData
+    internal abstract class StorageResourceCheckpointDetailsInternal : StorageResourceCheckpointDetails
     {
         internal void SerializeInternal(Stream stream) => Serialize(stream);
     }