+++ sdk/storage/Azure.Storage.DataMovement/tests/CleanUpTransferTests.cs
@@ -23,10 +23,10 @@ namespace Azure.Storage.DataMovement.Tests
                 .Returns("BlockBlob");
             mock.Setup(b => b.ProviderId)
                 .Returns("blob");
-            mock.Setup(b => b.GetSourceCheckpointData())
-                .Returns(new MockResourceCheckpointData());
-            mock.Setup(b => b.GetDestinationCheckpointData())
-                .Returns(new MockResourceCheckpointData());
+            mock.Setup(b => b.GetSourceCheckpointDetails())
+                .Returns(new MockResourceCheckpointDetails());
+            mock.Setup(b => b.GetDestinationCheckpointDetails())
+                .Returns(new MockResourceCheckpointDetails());
             mock.Setup(b => b.GetPropertiesAsync(It.IsAny<CancellationToken>()))
                 .Returns(Task.FromResult(new StorageResourceItemProperties()
                     {
@@ -52,10 +52,10 @@ namespace Azure.Storage.DataMovement.Tests
                 .Returns(Constants.GB);
             mock.Setup(b => b.MaxSupportedChunkSize)
                 .Returns(Constants.GB);
-            mock.Setup(b => b.GetSourceCheckpointData())
-                .Returns(new MockResourceCheckpointData());
-            mock.Setup(b => b.GetDestinationCheckpointData())
-                .Returns(new MockResourceCheckpointData());
+            mock.Setup(b => b.GetSourceCheckpointDetails())
+                .Returns(new MockResourceCheckpointDetails());
+            mock.Setup(b => b.GetDestinationCheckpointDetails())
+                .Returns(new MockResourceCheckpointDetails());
             mock.Setup(b => b.CompleteTransferAsync(It.IsAny<bool>(), It.IsAny<StorageResourceCompleteTransferOptions>(), It.IsAny<CancellationToken>()))
                 .Returns(Task.CompletedTask);
             // Throw a failure when doing a CopyFromUri call to trigger a failed state
@@ -88,7 +88,7 @@ namespace Azure.Storage.DataMovement.Tests
             source.Verify(b => b.ProviderId, Times.Once());
             source.Verify(b => b.ResourceId, Times.Once());
             source.Verify(b => b.Length, Times.Once());
-            source.Verify(b => b.GetSourceCheckpointData(), Times.Once());
+            source.Verify(b => b.GetSourceCheckpointDetails(), Times.Once());
             source.Verify(b => b.GetPropertiesAsync(It.IsAny<CancellationToken>()));
             source.Verify(b => b.GetCopyAuthorizationHeaderAsync(It.IsAny<CancellationToken>()));
             source.VerifyNoOtherCalls();
@@ -119,7 +119,7 @@ namespace Azure.Storage.DataMovement.Tests
             destMock.Verify(b => b.ResourceId, Times.Once());
             destMock.Verify(b => b.MaxSupportedSingleTransferSize, Times.Once());
             destMock.Verify(b => b.MaxSupportedChunkSize, Times.Once());
-            destMock.Verify(b => b.GetDestinationCheckpointData(), Times.Once());
+            destMock.Verify(b => b.GetDestinationCheckpointDetails(), Times.Once());
             destMock.Verify(b => b.SetPermissionsAsync(
                 sourceMock.Object,
                 It.IsAny<StorageResourceItemProperties>(),
@@ -163,7 +163,7 @@ namespace Azure.Storage.DataMovement.Tests
             destMock.Verify(b => b.ResourceId, Times.Once());
             destMock.Verify(b => b.MaxSupportedSingleTransferSize, Times.Once());
             destMock.Verify(b => b.MaxSupportedChunkSize, Times.Once());
-            destMock.Verify(b => b.GetDestinationCheckpointData(), Times.Once());
+            destMock.Verify(b => b.GetDestinationCheckpointDetails(), Times.Once());
             destMock.Verify(b => b.SetPermissionsAsync(
                 sourceMock.Object,
                 It.IsAny<StorageResourceItemProperties>(),